backend/requirements.txt
## Dependencias principales
Django>=5.2.7
djangorestframework>=3.15.1

## Autenticación / tokens / JWT
# DRF contiene rest_framework.authtoken; si usas JWT también está aquí:
djangorestframework-simplejwt>=5.3.1

## Bases de datos (MySQL y PostgreSQL helpers)
PyMySQL>=1.1.0      # usado para compatibilidad MySQL local (pymysql.install_as_MySQLdb())
psycopg2-binary>=2.9.11  # driver PostgreSQL (opcional según despliegue)

## CORS / Entorno / utilidades
django-cors-headers>=4.3.1
django-environ>=0.11.2

## Exportar a Excel
openpyxl>=3.1.5

## Internas / utilidades
sqlparse>=0.5.0
pytz>=2024.1
tzdata>=2024.1

## Opcionales / despliegue
python-dotenv>=1.0.0    # si prefieres .env en lugar de django-environ
gunicorn>=20.1.0        # servidor WSGI para producción
whitenoise>=6.5.0       # servir archivos estáticos en producción

## Compatibilidad .xls (antiguo)
# Si necesitas generar archivos .xls (Excel 97-2003) en lugar de .xlsx,
# instala `xlwt` y adapta los endpoints. Por defecto el proyecto usa .xlsx
# (openpyxl) que es el formato moderno y recomendado.
# xlwt>=1.3.0

# Notas:
# - El frontend tiene sus propias dependencias en frontend/package.json
# - Para instalar: `pip install -r requirements.txt`
